IdpIdTokenMappingService.java
