+++ src/main/java/com/icthh/xm/uaa/security/oauth2/idp/IdpTokenGranter.java
@@ -1,24 +1,18 @@
 package com.icthh.xm.uaa.security.oauth2.idp;
 
-import com.icthh.xm.commons.lep.LogicExtensionPoint;
 import com.icthh.xm.commons.lep.spring.LepService;
-import com.icthh.xm.commons.tenant.TenantContextUtils;
+import com.icthh.xm.commons.tenant.TenantContextHolder;
 import com.icthh.xm.uaa.domain.User;
-import com.icthh.xm.uaa.domain.UserLogin;
-import com.icthh.xm.uaa.domain.UserLoginType;
-import com.icthh.xm.uaa.domain.properties.TenantProperties;
 import com.icthh.xm.uaa.security.DomainUserDetails;
 import com.icthh.xm.uaa.security.DomainUserDetailsService;
 import com.icthh.xm.uaa.security.oauth2.idp.source.model.XmAuthenticationToken;
-import com.icthh.xm.uaa.service.TenantPropertiesService;
+import com.icthh.xm.uaa.service.IdpIdTokenMappingService;
 import com.icthh.xm.uaa.service.UserLoginService;
 import com.icthh.xm.uaa.service.UserService;
 import com.icthh.xm.uaa.service.dto.UserDTO;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.security.authentication.AuthenticationServiceException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.mapping.GrantedAuthoritiesMapper;
@@ -34,9 +28,10 @@ import org.springframework.security.oauth2.provider.token.AuthorizationServerTok
 
 import java.util.Collection;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+
+import static com.icthh.xm.uaa.security.DomainUserDetailsService.buildDomainUserDetails;
 
 /**
  * This class responsible for handling requests with idp token
@@ -47,16 +42,13 @@ public class IdpTokenGranter extends AbstractTokenGranter {
 
     public static final String GRANT_TYPE_IDP_TOKEN = "idp_token";
 
-    private static final String FIRST_NAME_CONFIG_ATTRIBUTE = "firstNameAttribute";
-    private static final String LAST_NAME_CONFIG_ATTRIBUTE = "lastNameAttribute";
-    private static final String USER_IDENTITY_CONFIG_TYPE = "userIdentityType";
-    private static final String USER_IDENTITY_CONFIG_ATTRIBUTE = "userIdentityAttribute";
-
     private final XmJwkTokenStore jwkTokenStore;
     private final DomainUserDetailsService domainUserDetailsService;
-    private final TenantPropertiesService tenantPropertiesService;
     private final UserService userService;
     private final UserLoginService userLoginService;
+    private final IdpIdTokenMappingService idpIdTokenMappingService;
+    private final TenantContextHolder tenantContextHolder;
+
     private final GrantedAuthoritiesMapper authoritiesMapper = new NullAuthoritiesMapper();
 
     public IdpTokenGranter(AuthorizationServerTokenServices tokenServices,
@@ -64,15 +56,17 @@ public class IdpTokenGranter extends AbstractTokenGranter {
                            @Lazy OAuth2RequestFactory requestFactory,
                            XmJwkTokenStore jwkTokenStore,
                            DomainUserDetailsService domainUserDetailsService,
-                           TenantPropertiesService tenantPropertiesService,
                            UserService userService,
-                           UserLoginService userLoginService) {
+                           UserLoginService userLoginService,
+                           IdpIdTokenMappingService idpIdTokenMappingService,
+                           TenantContextHolder tenantContextHolder) {
         super(tokenServices, clientDetailsService, requestFactory, GRANT_TYPE_IDP_TOKEN);
         this.jwkTokenStore = jwkTokenStore;
         this.domainUserDetailsService = domainUserDetailsService;
-        this.tenantPropertiesService = tenantPropertiesService;
         this.userService = userService;
         this.userLoginService = userLoginService;
+        this.idpIdTokenMappingService = idpIdTokenMappingService;
+        this.tenantContextHolder = tenantContextHolder;
     }
 
     @Override
@@ -99,7 +93,7 @@ public class IdpTokenGranter extends AbstractTokenGranter {
         // parse IDP id token
         OAuth2AccessToken idpOAuth2IdToken = jwkTokenStore.readAccessToken(idpToken);
 
-        validateIdpIdToken(idpOAuth2IdToken);
+        idpIdTokenMappingService.validateIdpIdToken(idpOAuth2IdToken);
 
         //user + role section
         DomainUserDetails userDetails = retrieveDomainUserDetails(idpOAuth2IdToken);
@@ -114,113 +108,26 @@ public class IdpTokenGranter extends AbstractTokenGranter {
     }
 
     private DomainUserDetails retrieveDomainUserDetails(OAuth2AccessToken idpOAuth2IdToken) {
-        String userIdentity = mapIdpIdTokenToIdentity(idpOAuth2IdToken);
+        String userIdentity = idpIdTokenMappingService.mapIdpIdTokenToIdentity(idpOAuth2IdToken);
         Optional<DomainUserDetails> userDetails = domainUserDetailsService.retrieveUserByUsername(userIdentity);
 
         if (userDetails.isEmpty()) {
             log.info("User not found by identity: {}, new user will be created", userIdentity);
             User newUser = createUser(userIdentity, idpOAuth2IdToken);
-            userDetails = Optional.of(DomainUserDetailsService.buildDomainUserDetails(userIdentity, getTenantKey(), newUser));
+            userDetails = Optional.of(buildDomainUserDetails(userIdentity, tenantContextHolder.getTenantKey(), newUser));
         }
         log.info("Mapped user for identity:{} is {}", userIdentity, userDetails);
         return userDetails.get();
     }
 
-    @LogicExtensionPoint(value = "MapIdpIdTokenToIdentity")
-    public String mapIdpIdTokenToIdentity(OAuth2AccessToken idpOAuth2IdToken) {
-        Map<String, Object> additionalInformation = idpOAuth2IdToken.getAdditionalInformation();
-        Map<String, String> defaultClaimsMapping = getDefaultClaimsMapping();
-
-        String userIdentityAttribute = defaultClaimsMapping.get(USER_IDENTITY_CONFIG_ATTRIBUTE);
-
-        return (String) additionalInformation.get(userIdentityAttribute);
-    }
-
     private User createUser(String userIdentity, OAuth2AccessToken idpOAuth2IdToken) {
-        UserDTO userDTO = mapIdpIdTokenToXmUser(userIdentity, idpOAuth2IdToken);
+        UserDTO userDTO = idpIdTokenMappingService.mapIdpIdTokenToXmUser(userIdentity, idpOAuth2IdToken);
 
         userLoginService.normalizeLogins(userDTO.getLogins());
         userLoginService.verifyLoginsNotExist(userDTO.getLogins());
 
-        userDTO.setRoleKey(mapIdpIdTokenToRole(userIdentity, idpOAuth2IdToken));
+        userDTO.setRoleKey(idpIdTokenMappingService.mapIdpIdTokenToRole(userIdentity, idpOAuth2IdToken));
 
         return userService.createUser(userDTO);
     }
-
-    @LogicExtensionPoint(value = "ValidateIdpIdToken")
-    public void validateIdpIdToken(OAuth2AccessToken idpOAuth2IdToken) {
-        //validate any additional claims viaLEP
-    }
-
-    @LogicExtensionPoint(value = "MapIdpIdTokenToXmUser")
-    public UserDTO mapIdpIdTokenToXmUser(String userIdentity, OAuth2AccessToken idpOAuth2IdToken) {
-        Map<String, Object> additionalInformation = idpOAuth2IdToken.getAdditionalInformation();
-        UserDTO userDTO = new UserDTO();
-
-        //base info mapping
-        Map<String, String> defaultClaimsMapping = getDefaultClaimsMapping();
-        String userFirstNameAttribute = defaultClaimsMapping.get(FIRST_NAME_CONFIG_ATTRIBUTE);
-
-        String userLastNameAttribute = defaultClaimsMapping.get(LAST_NAME_CONFIG_ATTRIBUTE);
-
-        userDTO.setFirstName((String) additionalInformation.get(userFirstNameAttribute));
-        userDTO.setLastName((String) additionalInformation.get(userLastNameAttribute));
-        //login mapping
-
-        String userIdentityType = UserLoginType
-            .fromString(defaultClaimsMapping.get(USER_IDENTITY_CONFIG_TYPE))
-            .getValue();
-
-        UserLogin emailUserLogin = new UserLogin();
-        emailUserLogin.setLogin(userIdentity);
-        emailUserLogin.setTypeKey(userIdentityType);
-
-        userDTO.setLogins(List.of(emailUserLogin));
-        return userDTO;
-    }
-
-    @SneakyThrows
-    @LogicExtensionPoint(value = "MapIdpIdTokenToRole")
-    public String mapIdpIdTokenToRole(String userIdentity, OAuth2AccessToken idpOAuth2IdToken) {
-        TenantProperties tenantProps = tenantPropertiesService.getTenantProps();
-        TenantProperties.Security security = tenantProps.getSecurity();
-
-        if (security == null || StringUtils.isEmpty(security.getDefaultUserRole())) {
-            log.debug("Default user role for tenant [{}] not specified in configuration.", getTenantKey());
-            throw new AuthenticationServiceException("Authentication failed " +
-                "cause of tenant [" + getTenantKey() + "] configuration lack.");
-        }
-        return security.getDefaultUserRole();
-    }
-
-    private String getTenantKey() {
-        return TenantContextUtils.getRequiredTenantKeyValue(tenantPropertiesService.getTenantContextHolder());
-    }
-
-    private Map<String, String> getDefaultClaimsMapping() {
-        TenantProperties tenantProps = tenantPropertiesService.getTenantProps();
-        TenantProperties.Security security = tenantProps.getSecurity();
-
-        TenantProperties.Security.Idp idp = security.getIdp();
-
-        if (idp == null || idp.getDefaultIdpClaimMapping() == null
-            || idp.getDefaultIdpClaimMapping().getFirstNameAttribute() == null
-            || idp.getDefaultIdpClaimMapping().getLastNameAttribute() == null
-            || idp.getDefaultIdpClaimMapping().getUserIdentityAttribute() == null
-            || idp.getDefaultIdpClaimMapping().getUserIdentityType() == null
-        ) {
-            log.debug("User Identity mapping not fully specified in tenant [{}] configuration.", getTenantKey());
-            throw new AuthenticationServiceException("Authentication failed " +
-                "cause of tenant [" + getTenantKey() + "] configuration lack.");
-        }
-        TenantProperties.Security.Idp.DefaultIdpClaimMapping defaultIdpClaimMapping = idp.getDefaultIdpClaimMapping();
-
-        return Map.of(
-            FIRST_NAME_CONFIG_ATTRIBUTE, defaultIdpClaimMapping.getFirstNameAttribute(),
-            LAST_NAME_CONFIG_ATTRIBUTE, defaultIdpClaimMapping.getLastNameAttribute(),
-            USER_IDENTITY_CONFIG_ATTRIBUTE, defaultIdpClaimMapping.getUserIdentityAttribute(),
-            USER_IDENTITY_CONFIG_TYPE, defaultIdpClaimMapping.getUserIdentityType()
-        );
-    }
-
 }
